requirements.txt
chromadb==0.6.3
pandas==2.1.4
numpy==1.26.4
pymysql==1.1.0
python-dotenv==1.0.0
openai>=1.0.0
mysql-connector-python==8.0.33
sentence-transformers>=2.6.0
huggingface_hub>=0.23.0
torch>=2.0.1
transformers>=4.30.2
openpyxl>=3.1.2
langchain-core>=0.3.41
langchain-openai>=0.3.8
langchain-community>=0.3.19
langchain-text-splitters>=0.3.6
langchain-huggingface>=0.1.2
pydantic>=2.8.2
faiss-cpu>=1.10.0
tensorflow>=2.12.0
tensorflow_hub>=0.12.0
tensorflow_text>=2.8.0
cryptography
streamlit>=1.31.0